% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_func.R
\name{beta_softmax_game}
\alias{beta_softmax_game}
\title{Title}
\usage{
beta_softmax_game(
  trial_num,
  pay1,
  pay2,
  temp,
  game_type = "blocked",
  cond_type = "mixed",
  aW = c(1, 1),
  bW = c(1, 1),
  gamma = 1,
  udb = 0,
  usb = 0,
  kf = 1
)
}
\arguments{
\item{trial_num}{number of trials per game}

\item{pay1}{payoff probability of the better arm}

\item{pay2}{payoff probability of the worse arm}

\item{temp}{inverse temperature of the sofmax choice function}

\item{game_type}{str: 'blocked' or 'er'. 'blocked' means mixed game choice sequence is blocked: 10 observe + 10 choose. 'er' (event-related) means mixed game choice sequence is randomized}

\item{cond_type}{str: 'mixed' or 'free' 'free':}

\item{aW}{2-num vector: weights for the alpha parameter (positive update)  (1st element for free-choice cond; 2nd <- for observe cond)}

\item{bW}{2-num vector: weight vector for the beta parameter (negative update)}
}
\value{
generates a single game using beta distribution to represent values
}
\description{
Title
}
