% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_path}
\alias{config_path}
\title{Retrieve the configurations path included with the package}
\usage{
config_path()
}
\value{
configuration path
}
\description{
Retrieve the configurations path included with the package
}
