% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{list_configurations}
\alias{list_configurations}
\title{List the configurations stored in the package (or some other place)}
\usage{
list_configurations(path = config_path(), ...)
}
\arguments{
\item{path}{chr the path to th configuration}

\item{...}{other argumentes for [base::list.files]}
}
\value{
character vector of zero or more configuration filenames
}
\description{
List the configurations stored in the package (or some other place)
}
