% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{obpg_path}
\alias{obpg_path}
\alias{set_obpg_path}
\alias{get_obpg_root}
\title{Get, set and build OBPG paths}
\usage{
set_obpg_path(root = ".", filename = "~/.obtools")

get_obpg_root(filename = "~/.obtools")

obpg_path(..., root = get_obpg_path())
}
\arguments{
\item{root}{chr the root OBPG data path}

\item{filename}{chr the filename where we store the path}

\item{...}{file path segments passed to [file.path]}
}
\value{
a file or directory path description
}
\description{
Get, set and build OBPG paths
}
