% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtools-package.R
\docType{package}
\name{obtools}
\alias{obtools-package}
\alias{obtools}
\title{Tools for use with raster data originating from the Ocean Biology Processing Group.}
\description{
Extra R-language tools to supplement [obpg R package](https://github.com/BigelowLab/obpg). 
This package serves up R scripts to update a local database on a regular basis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/obtools}
}

}
