% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{parse_param}
\alias{parse_param}
\title{Parse the "product" element of a config list to yield a new "param" element}
\usage{
parse_param(x)
}
\arguments{
\item{x}{configuration list}
}
\value{
the input list with "param" added if possible
}
\description{
Parse the "product" element of a config list to yield a new "param" element
}
