% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_configuration}
\alias{read_configuration}
\title{Read a configuration}
\usage{
read_configuration(
  filename = "nwa_AQUA_day.yaml",
  add_param = TRUE,
  path = config_path()
)
}
\arguments{
\item{filename}{chr the name of the file}

\item{path}{chr the path to th configuration}
}
\value{
configuration list
}
\description{
Read a configuration
}
